moving common VS projects settings into own property page

+++ ACM/ACMStream.cpp
@@ -387,9 +387,9 @@ unsigned int ACMStream::GetOutputSampleRate(int samples_per_sec, int bitrate, in
 	unsigned int OutputFrequency;
 	double compression_ratio = double(samples_per_sec * 16 * channels / (bitrate * 8));
 	if (compression_ratio > 13.)
-		OutputFrequency = map2MP3Frequency( (10. * bitrate * 8) / (16 * channels));
+		OutputFrequency = map2MP3Frequency( int((10. * bitrate * 8) / (16 * channels)));
 	else
-		OutputFrequency = map2MP3Frequency( 0.97 * samples_per_sec );
+		OutputFrequency = map2MP3Frequency( int(0.97 * samples_per_sec) );
 
 	return OutputFrequency;
